+++ partiql-value/src/datetime.rs
@@ -4,21 +4,25 @@ use std::cmp::Ordering;
 use std::fmt::{Debug, Formatter};
 use std::hash::Hash;
 use std::num::NonZeroU8;
+use time::macros::format_description;
 use time::{Duration, UtcOffset};
 
 #[derive(Hash, PartialEq, Eq, Clone)]
 #[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
 pub enum DateTime {
     Date(time::Date),
-    Time(time::Time),
-    TimeWithTz(time::Time, time::UtcOffset),
-    Timestamp(time::PrimitiveDateTime),
-    TimestampWithTz(time::OffsetDateTime),
+    Time(time::Time, Option<u32>),
+    TimeWithTz(time::Time, Option<u32>, time::UtcOffset),
+    Timestamp(time::PrimitiveDateTime, Option<u32>),
+    TimestampWithTz(time::OffsetDateTime, Option<u32>),
 }
 
 impl DateTime {
     pub fn from_hms(hour: u8, minute: u8, second: u8) -> Self {
-        DateTime::Time(time::Time::from_hms(hour, minute, second).expect("valid time value"))
+        DateTime::Time(
+            time::Time::from_hms(hour, minute, second).expect("valid time value"),
+            None,
+        )
     }
 
     pub fn from_hms_nano(hour: u8, minute: u8, second: u8, nanosecond: u32) -> Self {
@@ -63,11 +67,11 @@ impl DateTime {
         let date = time::Date::from_calendar_date(year, month, day).expect("valid ymd");
         let time = time_from_hms_nano(hour, minute, second, nanosecond);
         match offset {
-            None => DateTime::Timestamp(date.with_time(time)),
+            None => DateTime::Timestamp(date.with_time(time), None),
             Some(o) => {
                 let offset = UtcOffset::from_whole_seconds(o * 60).expect("offset in range");
                 let date = date.with_time(time).assume_offset(offset);
-                DateTime::TimestampWithTz(date)
+                DateTime::TimestampWithTz(date, None)
             }
         }
     }
@@ -81,9 +85,48 @@ impl DateTime {
     ) -> Self {
         let time = time_from_hms_nano(hour, minute, second, nanosecond);
         match offset {
-            Some(offset) => DateTime::TimeWithTz(time, offset),
-            None => DateTime::Time(time),
+            Some(offset) => DateTime::TimeWithTz(time, None, offset),
+            None => DateTime::Time(time, None),
         }
+    }
+
+    pub fn from_yyyy_mm_dd(date: &str) -> Self {
+        let format = format_description!("[year]-[month]-[day]");
+        let date = time::Date::parse(date, &format).expect("valid date string");
+        DateTime::Date(date)
+    }
+
+    pub fn from_hh_mm_ss(time: &str, precision: &Option<u32>) -> Self {
+        let format = format_description!("[hour]:[minute]:[second].[subsecond]");
+        let time = time::Time::parse(time, &format).expect("valid time string");
+        DateTime::Time(time, *precision)
+    }
+
+    pub fn from_hh_mm_ss_time_zone(time: &str, precision: &Option<u32>) -> Self {
+        let time_format = format_description!(
+            "[hour]:[minute]:[second].[subsecond][offset_hour]:[offset_minute]"
+        );
+        let time_part = time::Time::parse(time, &time_format).expect("valid time with time zone");
+        let time_format = format_description!(
+            "[hour]:[minute]:[second].[subsecond][offset_hour]:[offset_minute]"
+        );
+        let offset_part = time::UtcOffset::parse(time, &time_format).expect("valid time zone");
+        DateTime::TimeWithTz(time_part, *precision, offset_part)
+    }
+
+    pub fn from_yyyy_mm_dd_hh_mm_ss(timestamp: &str, precision: &Option<u32>) -> Self {
+        let format =
+            format_description!("[year]-[month]-[day] [hour]:[minute]:[second].[subsecond]");
+        let time =
+            time::PrimitiveDateTime::parse(timestamp, &format).expect("valid timestamp string");
+        DateTime::Timestamp(time, *precision)
+    }
+
+    pub fn from_yyyy_mm_dd_hh_mm_ss_time_zone(timestamp: &str, precision: &Option<u32>) -> Self {
+        let format = format_description!("[year]-[month]-[day] [hour]:[minute]:[second].[subsecond][offset_hour]:[offset_minute]");
+        let time = time::OffsetDateTime::parse(timestamp, &format)
+            .expect("valid timestamp string with time zone");
+        DateTime::TimestampWithTz(time, *precision)
     }
 }
 
@@ -97,18 +140,22 @@ impl Debug for DateTime {
             DateTime::Date(d) => {
                 write!(f, "DATE '{d:?}'")
             }
-            DateTime::Time(t) => {
-                write!(f, "TIME '{t:?}'")
-            }
-            DateTime::TimeWithTz(t, tz) => {
-                write!(f, "TIME WITH TIME ZONE '{t:?} {tz:?}'")
-            }
-            DateTime::Timestamp(dt) => {
-                write!(f, "TIMESTAMP '{dt:?}'")
-            }
-            DateTime::TimestampWithTz(dt) => {
-                write!(f, "TIMESTAMP WITH TIME ZONE '{dt:?}'")
-            }
+            DateTime::Time(t, p) => match p {
+                None => write!(f, "TIME '{t:?}'"),
+                Some(p) => write!(f, "TIME ({p:?}) '{t:?}'"),
+            },
+            DateTime::TimeWithTz(t, p, tz) => match p {
+                None => write!(f, "TIME WITH TIME ZONE '{t:?} {tz:?}'"),
+                Some(p) => write!(f, "TIME ({p:?}) WITH TIME ZONE '{t:?} {tz:?}'"),
+            },
+            DateTime::Timestamp(dt, p) => match p {
+                None => write!(f, "TIMESTAMP '{dt:?}'"),
+                Some(p) => write!(f, "TIMESTAMP ({p:?}) '{dt:?}'"),
+            },
+            DateTime::TimestampWithTz(dt, p) => match p {
+                None => write!(f, "TIMESTAMP WITH TIME ZONE '{dt:?}'"),
+                Some(p) => write!(f, "TIMESTAMP ({p:?}) WITH TIME ZONE '{dt:?}'"),
+            },
         }
     }
 }
@@ -127,25 +174,29 @@ impl Ord for DateTime {
             (DateTime::Date(_), _) => Ordering::Less,
             (_, DateTime::Date(_)) => Ordering::Greater,
 
-            (DateTime::Time(l), DateTime::Time(r)) => l.cmp(r),
-            (DateTime::Time(_), _) => Ordering::Less,
-            (_, DateTime::Time(_)) => Ordering::Greater,
+            (DateTime::Time(l, _lp), DateTime::Time(r, _rp)) => l.cmp(r),
+            // TODO: sorting using the time precisions
+            (DateTime::Time(_, _), _) => Ordering::Less,
+            (_, DateTime::Time(_, _)) => Ordering::Greater,
 
-            (DateTime::TimeWithTz(l, lo), DateTime::TimeWithTz(r, ro)) => {
+            (DateTime::TimeWithTz(l, _lp, lo), DateTime::TimeWithTz(r, _rp, ro)) => {
+                // TODO: sorting using the time precisions
                 let lod = Duration::new(lo.whole_seconds() as i64, 0);
                 let rod = Duration::new(ro.whole_seconds() as i64, 0);
                 let l_adjusted = *l + lod;
                 let r_adjusted = *r + rod;
                 l_adjusted.cmp(&r_adjusted)
             }
-            (DateTime::TimeWithTz(_, _), _) => Ordering::Less,
-            (_, DateTime::TimeWithTz(_, _)) => Ordering::Greater,
+            (DateTime::TimeWithTz(_, _, _), _) => Ordering::Less,
+            (_, DateTime::TimeWithTz(_, _, _)) => Ordering::Greater,
 
-            (DateTime::Timestamp(l), DateTime::Timestamp(r)) => l.cmp(r),
-            (DateTime::Timestamp(_), _) => Ordering::Less,
-            (_, DateTime::Timestamp(_)) => Ordering::Greater,
+            // TODO: sorting using the timestamp precisions
+            (DateTime::Timestamp(l, _lp), DateTime::Timestamp(r, _rp)) => l.cmp(r),
+            (DateTime::Timestamp(_, _), _) => Ordering::Less,
+            (_, DateTime::Timestamp(_, _)) => Ordering::Greater,
 
-            (DateTime::TimestampWithTz(l), DateTime::TimestampWithTz(r)) => l.cmp(r),
+            // TODO: sorting using the timestamp precisions
+            (DateTime::TimestampWithTz(l, _lp), DateTime::TimestampWithTz(r, _rp)) => l.cmp(r),
         }
     }
 }